+++ src/main.rs
@@ -1,81 +1,15 @@
 use anyhow::Result;
 use clap::Parser;
-use std::{env, fs, io::Write};
+use std::{env, fs::create_dir, io::Write};
 
+mod benchmarking;
+mod cli;
 mod config;
 mod dependency;
-mod generate;
-mod run;
+mod project;
 mod subprocess;
 
-use crate::{
-    config::Config,
-    dependency::{parse_dependency, Dependency},
-    run::execute,
-};
-
-/// This tool allow you to create a new Rust temporary project in a temporary
-/// directory.
-///
-/// The dependencies can be provided in arguments (e.g.`cargo-temp anyhow
-/// tokio`). When the shell is exited, the temporary directory is deleted unless
-/// you removed the file `TO_DELETE`.
-#[derive(Parser, Debug, Clone)]
-#[command(author, version, about, long_about)]
-pub struct Cli {
-    /// Dependencies to add to `Cargo.toml`.
-    ///
-    /// The default version used is `*` but this can be replaced using `=`.
-    /// E.g. `cargo-temp anyhow=1.0.13`
-    #[arg(value_parser = parse_dependency)]
-    pub dependencies: Vec<Dependency>,
-
-    /// Create a library instead of a binary.
-    #[arg(long, short = 'l')]
-    pub lib: bool,
-
-    /// Name of the temporary crate.
-    #[arg(long = "name", short = 'n')]
-    pub project_name: Option<String>,
-
-    /// Create a temporary Git working tree based on the repository in the
-    /// current directory.
-    #[arg(long = "worktree", short = 'w')]
-    pub worktree_branch: Option<Option<String>>,
-
-    /// Create a temporary clone of a Git repository.
-    #[arg(long, short = 'g')]
-    pub git: Option<String>,
-
-    /// Add a `benches` to the temporary project.
-    ///
-    /// You can choose the name of the benchmark file name as argument.
-    /// The default is `benchmark.rs`
-    #[arg(long, short = 'b')]
-    pub bench: Option<Option<String>>,
-
-    /// Select the Rust's edition of the temporary project.
-    ///
-    /// Available options are:
-    /// * 15 | 2015 => edition 2015,
-    /// * 18 | 2018 => edition 2018,
-    /// * 21 | 2021 => edition 2021,
-    ///
-    /// If the argument doesn't match any of the options, the default is the latest edition
-    #[arg(long, short = 'e')]
-    pub edition: Option<u32>,
-
-    #[cfg(feature = "generate")]
-    #[command(subcommand)]
-    pub subcommand: Option<Subcommand>,
-}
-
-#[derive(Clone, Debug, clap::Subcommand)]
-pub enum Subcommand {
-    /// Generate a temporary project from a template using `cargo-generate`.
-    #[cfg(feature = "generate")]
-    Generate(generate::Args),
-}
+use crate::{cli::Cli, config::Config, project::Project};
 
 fn main() -> Result<()> {
     env_logger::builder()
@@ -92,27 +26,7 @@ fn main() -> Result<()> {
 
     // Read configuration from disk or generate a default one.
     let config = Config::get_or_create()?;
-    let _ = fs::create_dir(&config.temporary_project_dir);
+    let _ = create_dir(&config.temporary_project_dir);
 
-    #[cfg(feature = "generate")]
-    match cli.subcommand {
-        Some(Subcommand::Generate(args)) => args.generate(config)?,
-        None => execute(cli, config)?,
-    }
-
-    #[cfg(not(feature = "generate"))]
-    execute(cli, config)?;
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_cli() {
-        use clap::CommandFactory;
-        Cli::command().debug_assert()
-    }
+    Project::execute(cli, config)
 }
